\docType{package}
\name{robustmeta-package}
\alias{robustmeta-package}
\title{The 'robustmeta' package.}
\description{
A R package for implementing the robust inference methods for meta-analysis involving influential outlying studies.
}
\references{
Noma, H., Sugasawa, S. and Furukawa, T. A. (2022).
Robust inference methods for meta-analysis involving influential outlying studies. In Preparation.
}
