% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{varenicline}
\alias{varenicline}
\title{Thomas et al. (2015)'s varenicline data}
\format{A data frame with 29 rows and 5 variables}
\usage{
data(varenicline)
}
\description{
\itemize{
  \item \code{study}: Study ID
  \item \code{d1}: Number of depression events in treatment group
  \item \code{n1}: Number of observations in treatment group
  \item \code{d0}: Number of depression events in control group
  \item \code{n0}: Number of observations in control group
}
}
\references{
Thomas, K. H., Martin, R. M., Knipe, D. W., Higgins, J. P., Gunnell, D. (2015).
Risk of neuropsychiatric adverse events associated with varenicline: systematic review and meta-analysis.
\emph{BMJ.}
\strong{350}: h1109.
}
\keyword{datasets}
