\docType{data}
\name{broad}
\alias{broad}
\title{Broad vs. narrow selectivity data for robustmeta package. Based on
in-progress work with a NESCent working group. Not to be considered
final data.}
\description{
  Broad vs. narrow selectivity data for robustmeta package.
  Based on in-progress work with a NESCent working group.
  Not to be considered final data.
}
\keyword{data}
\keyword{datasets}

